fix(docs-site): render document metadata in the Html wrapper, not the Page
Closes the head-duplication follow-up (bd ea5): the docs `<head>` had two of
everything (title, charset, description, style) after hydration — head 5 -> 9.

Root cause (not a build divergence — both streams share the Page by design):
document metadata lived inside the Page, which is the tree the client hydrates.
React hoisted it into the prerendered <head> on the server, then re-hoisted it
on client mount (it doesn't dedup against head nodes it didn't author this
session). This predated all hydration work — the metadata has been in the Page
since the first docs commit (de409c4, #111).

Fix: move every document-level concern into a custom Html wrapper (html.tsx),
which vprs renders for the FULL (HTML-generating) stream only — the headless
stream the browser hydrates is page-only. So metadata lands in the prerendered
<head> exactly once and is never part of the hydrated tree.

- html.tsx: new Html wrapper — charset/viewport/<title>/description/<style>.
  Per-route <title> comes from pageProps.title.
- docs.ts: extracted the data layer (find/list/load docs, resolveTitle) out of
  page.tsx so props.ts and the Html wrapper can resolve a route's title without
  pulling in markdown rendering.
- style.ts: the site stylesheet, inlined by the Html <head>.
- page.tsx: dropped the metadata block, STYLE, and title computation; now pure
  render layer importing the data layer.
- props.ts: provides title via resolveTitle(url).
- vite.config.ts: registers Html: src/page/html.tsx.
- client.tsx: per-route document.title on client nav, set imperatively (no
  hoisting, no duplicate <title>).

Verified (headless, prod build): prerendered <head> has title/charset/style
once; after hydration headChildren 6 (was 9), exactly one <title>/<style>, no
#418, zero .rsc on load, correct per-route title on load AND nav, Shiki
highlighting + CSS intact.
