+++ docs-site/src/client.tsx
@@ -70,6 +70,16 @@ function App({ initial }: { initial: RscNode }) {
     };
   }, []);
 
+  // Keep the document <title> in sync on client navigation. The Html wrapper
+  // renders the correct per-route title at prerender (and on full loads); after
+  // an in-site nav only #root is swapped, so update it here from the rendered
+  // <h1>. Imperative (document.title), so nothing is hoisted into <head> — no
+  // duplicate <title> on hydration.
+  useEffect(() => {
+    const h1 = document.querySelector("main.doc h1")?.textContent?.trim();
+    if (h1) document.title = `${h1} — vite-plugin-react-server`;
+  }, [content]);
+
   return use(content);
 }
