+++ docs-site/src/page/html.tsx
@@ -1,0 +1,56 @@
+import React from "react";
+import {
+  Css,
+  type HtmlComponentType,
+  type HtmlProps,
+} from "vite-plugin-react-server/components";
+import { STYLE } from "./style.js";
+
+/**
+ * The docs site's document shell.
+ *
+ * vprs renders this for the FULL (HTML-generating) stream only — the headless
+ * stream the browser hydrates is page-only (no <html>/<head>/<body>). So every
+ * document-level concern lives here: it lands in the prerendered <head> exactly
+ * once and is never part of the tree React hydrates, which is what keeps React
+ * from re-hoisting it on the client and duplicating the <head>.
+ *
+ * The per-route <title> comes from props.ts (pageProps.title); client-side
+ * navigation updates it imperatively in client.tsx.
+ */
+export const Html: HtmlComponentType<any, any, any, any> = ({
+  Root,
+  cssFiles,
+  globalCss,
+  pageProps,
+  Page,
+  as = "div",
+}: HtmlProps) => {
+  const title =
+    (pageProps as { title?: string } | undefined)?.title ??
+    "vite-plugin-react-server";
+  return (
+    <html lang="en">
+      <head>
+        <meta charSet="utf-8" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <title>{title}</title>
+        <meta
+          name="description"
+          content="React Server Components for Vite on stable React — no framework, no experimental channel. These docs are statically generated by the plugin itself."
+        />
+        <style>{STYLE}</style>
+        <Css cssFiles={globalCss} />
+      </head>
+      <body>
+        <Root
+          as={as}
+          id="root"
+          cssFiles={cssFiles}
+          pageProps={pageProps}
+          Page={Page}
+        />
+      </body>
+    </html>
+  );
+};
