+++ docs-site/src/page/props.ts
@@ -1,2 +1,7 @@
-/** Per-route props: the page server component picks its markdown by url. */
-export const props = (url: string) => ({ url });
+import { resolveTitle } from "./docs.js";
+
+/**
+ * Per-route props: the page picks its markdown by url; `title` feeds the
+ * document <title> rendered by the Html wrapper (html.tsx).
+ */
+export const props = (url: string) => ({ url, title: resolveTitle(url) });
